requirements.txt
APScheduler==3.10.1
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.2
Babel==2.12.1
black==23.3.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dacite==1.8.0
decorator==4.4.2
dnspython==2.3.0
ffmpeg-python==0.2.0
Flask==2.2.2
Flask-Cors==3.0.10
future==0.18.3
google-api-core==2.11.0
google-auth==2.16.2
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
gunicorn==20.1.0
idna==3.4
imageio==2.26.0
imageio-ffmpeg==0.4.8
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
moviepy==1.0.3
mypy-extensions==1.0.0
numpy==1.24.2
packaging==23.1
pathspec==0.11.1
pdfkit==1.0.0
Pillow==9.4.0
platformdirs==3.5.1
proglog==0.1.10
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
PyJWT==2.6.0
pymongo==4.3.2
pytz==2023.3
redis==4.5.4
requests==2.28.2
rq==1.13.0
rsa==4.9
six==1.16.0
tomli==2.0.1
tqdm==4.65.0
tzlocal==5.0
urllib3==1.26.15
Werkzeug==2.2.2
